msm: vidc: fix vidc kernel memory leakage for secure vp9 decoding
For secure vp9 decoding case, it will add timestamp node when queue
buffer in msm_vidc_qbuf(), but it won't be removed because is_valid
of timestamp node is not set to false when dequeue buffer.
Add case to call msm_comm_fetch_ts_framerate api, then is_valid is
set to false.

Change-Id: Ib885e042aac92863de0d6ee7c8ca62a3938fb259
Signed-off-by: chlei <quic_chlei@quicinc.com>